PLAYNEXT-8904 Improve migration view
